# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nexus Pro Backend API
  version: 0.2.0
  description: |
    Identifiers for persisted tenant resources are globally unique across tenants.
    tenant_id scopes ownership, authorization, and row-level security.
security:
  - bearerAuth: []
paths:
...
  /v1/forms/templates:
    get:
      summary: List form templates
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: Form template page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
    post:
      summary: Create form template
      responses:
        "201":
          description: Created form template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
  /v1/workflows/forms:
    get:
      summary: List workflow form instances
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Sort"
        - {name: status, in: query, schema: {type: string, enum: [all, submitted, pending, approved, rejected, cancelled]}}
        - {name: template_id, in: query, schema: {type: string}}
        - {name: template_key, in: query, schema: {type: string}}
        - {name: applicant_account_id, in: query, schema: {type: string}}
        - {name: mine, in: query, schema: {type: boolean}}
      responses:
        "200":
          description: Workflow form instance page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workflows/reviews:
    get:
      summary: Group workflow form instances for review notification tabs
      responses:
        "200":
          description: Workflow review queue projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workflows/reviews/bulk-action:
    post:
      summary: Apply one workflow review action to multiple form instances
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkReviewFormsRequest"
      responses:
        "200":
          description: Bulk workflow review result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "207":
          description: Bulk workflow review result with per-item failures
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workflows/forms/drafts:
    post:
      summary: Save a workflow form draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveFormDraftRequest"
      responses:
        "201":
          description: Created workflow form draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workflows/forms/{id}:
    patch:
      summary: Update a workflow form draft
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFormDraftRequest"
      responses:
        "200":
          description: Updated workflow form draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
    delete:
      summary: Delete a workflow form draft
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Deleted workflow form draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workflows/forms/{id}/export:
    get:
      summary: Download a workflow form instance as a JSON file
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Workflow form JSON file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workflows/forms/{id}/submit:
    post:
      summary: Submit a workflow form
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFormRequest"
      responses:
        "201":
          description: Submitted workflow form instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workflows/forms/{id}/approve:
    post:
      summary: Approve the current business workflow form stage
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: {type: string}
      responses:
        "200":
          description: Approved workflow form instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workflows/forms/{id}/reject:
    post:
      summary: Reject a workflow form instance
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: {type: string}
      responses:
        "200":
          description: Rejected workflow form instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workflows/forms/{id}/return:
    post:
      summary: Return a workflow form instance for applicant revision
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: {type: string}
      responses:
        "200":
          description: Returned workflow form instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workflows/forms/{id}/cancel:
    post:
      summary: Cancel a workflow form instance
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelFormRequest"
      responses:
        "200":
          description: Cancelled workflow form instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workflows/forms/{id}/workflow:
    get:
      summary: Get workflow runtime state for a form instance
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Workflow runtime state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowFormStateDataResponse"
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/workflows/forms/{id}/duplicate:
    post:
      summary: Duplicate a workflow form instance into a new draft
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "201":
          description: Created duplicate workflow form draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token issued by the configured identity provider and mapped to a local tenant account.
  parameters:
    ID:
      name: id
      in: path
      required: true
      description: Globally unique persisted resource identifier; tenant_id still scopes ownership and authorization.
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
    Sort:
      name: sort
      in: query
      schema:
        type: string
...
  responses:
    ValidationError:
      description: Validation failed. Field errors include tab, field, code and message; import row errors include row_number.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthenticated:
      description: Authenticated tenant/account context is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Permission, menu, button, field, or data-scope access was denied.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource was not found in the current tenant.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Request conflicts with current resource state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error. Response includes trace_id.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    DataResponse:
      type: object
      required: [data]
      properties:
        data:
          description: Successful response payload.
...
    WorkflowFormStateDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/WorkflowFormStateResponse"
    WorkflowFormStateResponse:
      type: object
      required: [form_instance_id, can_act, steps]
      properties:
        form_instance_id: {type: string}
        run_id: {type: string}
        run_status:
          type: string
          enum: [running, returned, completed, cancelled]
        current_stage_id: {type: string}
        current_stage_label: {type: string}
        can_act: {type: boolean}
        allowed_actions:
          type: array
          items:
            type: string
            enum: [approve, reject, return]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowFormStep"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowReviewLogItem"
    WorkflowFormStep:
      type: object
      required: [stage_id, label, state]
      properties:
        stage_id: {type: string}
        label: {type: string}
        detail: {type: string}
        state:
          type: string
          enum: [pending, active, completed, skipped, rejected]
        assignees:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowFormStepAssignee"
    WorkflowFormStepAssignee:
      type: object
      required: [account_id, name, status]
      properties:
        account_id: {type: string}
        name: {type: string}
        status:
          type: string
          enum: [pending, approved, rejected, returned]
    WorkflowReviewLogItem:
      type: object
      required: [type, name, time]
      properties:
        type: {type: string}
        name: {type: string}
        role: {type: string}
        time: {type: string, format: date-time}
        comment: {type: string}
...
    BulkReviewFormsRequest:
      type: object
      required: [form_instance_ids, action]
      properties:
        form_instance_ids:
          type: array
          minItems: 1
          items: {type: string}
        action:
          type: string
          enum: [approve, reject, return]
        reason: {type: string}
    SaveFormDraftRequest:
      type: object
      required: [template_key]
      properties:
        template_key: {type: string}
        payload:
          type: object
          additionalProperties: true
    UpdateFormDraftRequest:
      type: object
      properties:
        template_key:
          type: string
          description: Optional template key when changing the draft template.
        payload:
          type: object
          additionalProperties: true
    SubmitFormRequest:
      type: object
      properties:
        template_key:
          type: string
          description: Optional when the template key is supplied by the path id.
        payload:
          type: object
          additionalProperties: true
    CancelFormRequest:
      type: object
      properties:
        reason: {type: string}
...
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        tab:
          type: string
          description: Employee modal tab or section key, for example basic_info or employment_info.
        field:
          type: string
          description: Request or spreadsheet field key that failed validation.
        code:
          type: string
          description: Stable machine-readable validation or field-policy code.
        message:
          type: string
    RowError:
      type: object
      required: [row_number, field_errors]
      properties:
        row_number:
          type: integer
          minimum: 1
          description: Spreadsheet or CSV row number, including the header row offset used by the import parser.
        field_errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id]
          properties:
            code:
              type: integer
              format: int32
              description: |
                Numeric application error code. Current ranges:
                1xxxx common platform, request and authentication errors;
                2xxxx IAM and authorization errors; 3xxxx people-domain and HR errors;
                4xxxx attendance errors; 5xxxx workflow errors; 6xxxx agent errors.
              enum:
                - 10000
                - 10001
                - 10002
                - 10003
                - 10004
                - 10005
                - 10006
                - 10007
                - 10008
                - 10009
                - 10030
                - 10031
                - 10032
                - 10033
                - 10034
                - 10035
                - 10036
                - 10037
                - 10050
                - 10060
                - 10070
                - 20040
                - 20041
                - 20042
                - 20043
                - 20044
                - 20045
                - 30010
                - 30011
                - 30012
                - 30013
                - 30014
                - 30015
                - 30016
                - 30017
                - 30018
                - 30030
                - 30031
                - 30040
                - 30041
                - 30042
            message: {type: string}
            reason_code:
              type: string
              enum:
                - account_inactive
                - sso_email_not_authorized
                - sso_email_unverified
                - company_inactive
                - google_login_failed
                - sso_identity_conflict
                - sso_email_ambiguous
                - permission_missing
                - menu_denied
                - button_denied
                - field_denied
                - data_scope_denied
            field_errors:
              type: array
              items:
                $ref: "#/components/schemas/FieldError"
            row_errors:
              type: array
              items:
                $ref: "#/components/schemas/RowError"
            trace_id: {type: string}